+++ mirai-core/src/commonMain/kotlin/contact/info/FriendGroupInfo.kt
